+++ src/bin/commands/aux-create.ts
@@ -1,16 +1,18 @@
 #! /usr/bin/env node
 
-require('dotenv').config()
-const { create } = require('../../src/migrator')
-const { utcTimestamp } = require('../../lib/date')
-const log = require('../../lib/log')
+import * as dotenv from 'dotenv'
+import * as log from '../../lib/log'
+import { create } from '../../migrator'
+import { utcTimestamp } from '../../lib/date'
 
-exports.command = 'aux:create [name]'
+dotenv.config()
 
-exports.desc =
+export const command = 'aux:create [name]'
+
+export const desc =
   'Creates an auxiliary environment based on CTF_ENVIRONMENT_ID and applies newer migrations to it.'
 
-exports.builder = (yargs) => {
+export const builder = (yargs) => {
   yargs.positional('name', {
     describe: 'Specify an optional name for the auxiliary environment.',
     type: 'string',
@@ -18,7 +20,7 @@ exports.builder = (yargs) => {
   })
 }
 
-exports.handler = async ({ name }) => {
+export const handler = async ({ name }) => {
   try {
     await create({ newEnvId: name })
   } catch (e) {