+++ src/seeder.ts
@@ -1,7 +1,7 @@
-const spaceModule = require('../lib/contentful-space-manager')
+import CmpSpace from './lib/contentful-space-manager'
 
 const seed = async ({ spaceId, environmentId, accessToken }, type, data) => {
-  const space = await spaceModule(spaceId, environmentId, accessToken)
+  const space = await CmpSpace(spaceId, environmentId, accessToken)
   await space.createEntries(type, data)
 }